move tx building to v3 builder
