+++ src/config/v3TokenBuilder.ts
@@ -1,3 +1,8 @@
+import { encodeFunctionData, isAddress, stringify } from 'viem';
+import { Clanker_v3_1_abi } from '../abi/v3.1/Clanker.js';
+import { CLANKER_FACTORY_V3_1, DEFAULT_SUPPLY } from '../constants.js';
+import { getTokenPairByAddress } from '../services/desiredPrice.js';
+import { findVanityAddress } from '../services/vanityAddress.js';
 import type {
   ClankerMetadata,
   ClankerSocialContext,
@@ -6,6 +11,8 @@ import type {
   TokenConfig,
   VaultConfig,
 } from '../types/index.js';
+import { getDesiredPriceAndPairAddress } from '../utils/desired-price.js';
+import { getRelativeUnixTimestamp } from '../utils/unix-timestamp.js';
 
 /**
  * Builder class for creating TokenConfig objects
@@ -104,5 +111,162 @@ export class TokenConfigV3Builder {
       throw new Error('Token name and symbol are required');
     }
     return this.config as TokenConfig;
+  }
+
+  async buildTransaction(requestorAddress: `0x${string}`, chainId: number, config?: TokenConfig) {
+    const cfg = config ?? this.build();
+
+    if (!cfg.name || !cfg.symbol) {
+      throw new Error('Token name and symbol are required');
+    }
+
+    const poolConfig = {
+      quoteToken: cfg.pool?.quoteToken || '0x4200000000000000000000000000000000000006', // Default to WETH
+      initialMarketCap: cfg.pool?.initialMarketCap || '10', // Default to 10 ETH
+    };
+
+    const { desiredPrice, pairAddress } = getDesiredPriceAndPairAddress(
+      getTokenPairByAddress(poolConfig.quoteToken as `0x${string}`),
+      poolConfig.initialMarketCap
+    );
+
+    // Calculate price tick using the same method as the example
+    const logBase = 1.0001;
+    const tickSpacing = 200;
+    console.log('desiredPrice', desiredPrice);
+    const rawTick = Math.log(desiredPrice) / Math.log(logBase);
+    const initialTick = Math.floor(rawTick / tickSpacing) * tickSpacing;
+    console.log('initialTick', initialTick);
+
+    // Extract social media links with safe defaults
+    const socialLinks = cfg.metadata?.socialMediaUrls ?? [];
+    const telegramLink = socialLinks.find((url) => url.includes('t.me')) || '';
+    const xLink =
+      socialLinks.find((url) => url.includes('twitter.com') || url.includes('x.com')) || '';
+    const websiteLink =
+      socialLinks.find(
+        (url) => !url.includes('t.me') && !url.includes('twitter.com') && !url.includes('x.com')
+      ) || '';
+    const farcasterLink =
+      socialLinks.find((url) => url.includes('warpcast.com') || url.includes('farcaster.com')) ||
+      '';
+
+    // Build metadata with social links like in the example
+    const metadata = stringify({
+      description: cfg.metadata?.description || '',
+      auditUrls: cfg.metadata?.auditUrls ?? [],
+      socialMediaUrls: [
+        ...socialLinks,
+        ...(telegramLink ? [{ platform: 'telegram', url: telegramLink }] : []),
+        ...(websiteLink ? [{ platform: 'website', url: websiteLink }] : []),
+        ...(xLink ? [{ platform: 'x', url: xLink }] : []),
+        ...(farcasterLink ? [{ platform: 'farcaster', url: farcasterLink }] : []),
+      ],
+    });
+
+    const socialContext = stringify({
+      interface: cfg.context?.interface || 'SDK',
+      platform: cfg.context?.platform || '',
+      messageId: cfg.context?.messageId || '',
+      id: cfg.context?.id || '',
+    });
+
+    // Validate all addresses to make sure they're not empty or undefined
+    const validateAddress = (
+      address: string | undefined,
+      defaultAddress: `0x${string}`
+    ): `0x${string}` => {
+      if (!address || !isAddress(address)) {
+        return defaultAddress;
+      }
+      return address;
+    };
+
+    const admin = validateAddress(cfg.rewardsConfig?.creatorAdmin, requestorAddress);
+    const { token: expectedAddress, salt } = await findVanityAddress(
+      [
+        cfg.name,
+        cfg.symbol,
+        DEFAULT_SUPPLY,
+        admin,
+        cfg.image || '',
+        metadata,
+        socialContext,
+        BigInt(chainId),
+      ],
+      admin,
+      '0x4b07',
+      {
+        chainId: chainId,
+      }
+    );
+
+    const vestingUnlockDate = cfg.vault?.durationInDays
+      ? BigInt(Math.floor(Date.now() / 1000) + cfg.vault.durationInDays * 24 * 60 * 60)
+      : BigInt(0);
+
+    // Convert absolute timestamp to duration if provided
+    const vestingDuration = vestingUnlockDate
+      ? getRelativeUnixTimestamp(Number(vestingUnlockDate))
+      : BigInt(0);
+
+    const tokenConfig = {
+      tokenConfig: {
+        name: cfg.name,
+        symbol: cfg.symbol,
+        salt: salt,
+        image: cfg.image || '',
+        metadata: metadata,
+        context: socialContext,
+        originatingChainId: BigInt(chainId),
+      },
+      poolConfig: {
+        pairedToken: pairAddress,
+        tickIfToken0IsNewToken: initialTick || 0,
+      },
+      initialBuyConfig: {
+        pairedTokenPoolFee: 10000,
+        pairedTokenSwapAmountOutMinimum: 0n,
+      },
+      vaultConfig: {
+        vaultDuration: vestingDuration,
+        vaultPercentage: cfg.vault?.percentage || 0,
+      },
+      rewardsConfig: {
+        creatorReward: BigInt(cfg.rewardsConfig?.creatorReward || 40),
+        creatorAdmin: validateAddress(cfg.rewardsConfig?.creatorAdmin, requestorAddress),
+        creatorRewardRecipient: validateAddress(
+          cfg.rewardsConfig?.creatorRewardRecipient,
+          requestorAddress
+        ),
+        interfaceAdmin: validateAddress(cfg.rewardsConfig?.interfaceAdmin, requestorAddress),
+        interfaceRewardRecipient: validateAddress(
+          cfg.rewardsConfig?.interfaceRewardRecipient,
+          requestorAddress
+        ),
+      },
+    } as const;
+
+    try {
+      const deployCalldata = encodeFunctionData({
+        abi: Clanker_v3_1_abi,
+        functionName: 'deployToken',
+        args: [tokenConfig],
+      });
+
+      return {
+        transaction: {
+          to: CLANKER_FACTORY_V3_1,
+          data: deployCalldata,
+        },
+        expectedAddress,
+      };
+    } catch (error: unknown) {
+      console.error('Error encoding function data:', error);
+      console.error('Problematic deployArgs:', tokenConfig);
+      throw new Error(
+        `Failed to encode function data: ${error instanceof Error ? error.message : error}`
+      );
+    }
   }
 }
